requirements.txt
Django==3.2.12
djangorestframework==3.13.1
djangorestframework-simplejwt==4.6.0